+++ Shared/Sources/DesignSystem/Sources/PretendardDolDamFonts.swift
@@ -1,0 +1,33 @@
+//
+//  PretendardDolDamFonts.swift
+//
+//
+//  Created by 홍승현 on 1/30/24.
+//
+
+import Foundation
+
+// MARK: - Pretendard
+
+enum Pretendard: String, CaseIterable {
+  case black = "Pretendard-Black"
+
+  case extraBold = "Pretendard-ExtraBold"
+  case bold = "Pretendard-Bold"
+  case semiBold = "Pretendard-SemiBold"
+
+  case medium = "Pretendard-Medium"
+
+  case regular = "Pretendard-Regular"
+
+  case light = "Pretendard-Light"
+  case extraLight = "Pretendard-ExtraLight"
+
+  case thin = "Pretendard-Thin"
+}
+
+// MARK: - DolDam
+
+enum DolDam: String, CaseIterable {
+  case `default` = "EF_jejudoldam"
+}
